+++ packages/runtime/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./RuntimeEvent";
 export * from "./RuntimeHost";
+export * from "./RuntimeModule";
 export * from "./RuntimeServiceKeys";